+++ tests/testthat/test.transformations.R
@@ -49,3 +49,56 @@ test_that("Header transformation works", {
   expect_identical(res, expect)
 })
 
+test_that("disaggregate", {
+  # Correct number of rows.
+  dt <- fread("
+    sku weight
+    CUSISLA03BLU-UK 3
+    RUGISIS04BEI-UK 4
+    TBLBRM001WHI-UK 0",
+    header = T,
+    stringsAsFactors = F
+  )
+
+  calculated <- disaggregate(dt = dt, by = "weight")
+  expected <- fread("
+    sku weight
+    CUSISLA03BLU-UK 1
+    CUSISLA03BLU-UK 1
+    CUSISLA03BLU-UK 1
+    RUGISIS04BEI-UK 1
+    RUGISIS04BEI-UK 1
+    RUGISIS04BEI-UK 1
+    RUGISIS04BEI-UK 1",
+    header = T,
+    stringsAsFactors = F
+  )
+  expect_equal(calculated, expected,
+               label = "Each line is split into multiple matching the weight.")
+
+  # Data table retains its class.
+  dt <- fread("
+    sku weight
+    CUSISLA03BLU-UK 3
+    RUGISIS04BEI-UK 4
+    TBLBRM001WHI-UK 0",
+    header = TRUE,
+    stringsAsFactors = FALSE
+  )
+  calculated <- disaggregate(dt = dt, by = "weight")
+  expect_equal(sum(calculated$weight), sum(dt$weight),
+               label = "The total weight matches that of the original.")
+
+  # Row numbers are kept correctly.
+  dt <- fread("
+    sku weight
+    CUSISLA03BLU-UK 3
+    RUGISIS04BEI-UK 0
+    TBLBRM001WHI-UK 2",
+    header = TRUE,
+    stringsAsFactors = FALSE
+  )
+  calculated <- disaggregate(dt = dt, by = "weight", keep.row.number = TRUE)
+  expect_true(all(calculated$rn, as.integer(rep(rownames(dt), dt$weight))),
+               label = "The total weight matches that of the original.")
+})