+++ news.md
@@ -3,6 +3,7 @@
 * `s3GetFile()`, `s3PutFile()` - functions are changed to read data based on the extention of the file. Fore example you can use it instead of calling `s3GetFile.csv()` if path ends with `.csv`.
 
 * `bqInsertData()` - added fields parameter to force BigQuery types to the given types
+* `disaggregate()` - new function to split data.table from aggregated to individual lines
 
 # retl 0.1.1